Remove S3FileSystemFactory.
