+++ src/test/resources/application.properties
@@ -10,12 +10,12 @@ inbox.cache.ttl=3600
 
 inbox.local.directory=#{systemProperties['user.dir']}
 
+inbox.s3.region=${INBOX_S3_REGION:us-east-1}
 inbox.s3.host=${INBOX_S3_HOST:s3}
 inbox.s3.port=${INBOX_S3_PORT:9000}
 inbox.s3.access-key=${S3_ACCESS_KEY:}
 inbox.s3.secret-key=${S3_SECRET_KEY:}
 inbox.s3.use-ssl=${USE_SSL:false}
-inbox.s3.path-style-access=${PATH_STYLE_ACCESS:true}
 
 inbox.cega.endpoint=${CEGA_ENDPOINT:http://localhost:9100/lega/v1/legas/users/%s?idType=username}
 inbox.cega.credentials=lega:B2EannpNMdNuX8Rd
